// src/api/orval/model/getApiOpenVpnClientsGetAllConnectedParams.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */

export type GetApiOpenVpnClientsGetAllConnectedParams = {
  /**
   * @minimum 1
   * @maximum 2147483647
   */
  VpnServerId: number;
  /**
   * @minimum 1
   * @maximum 2147483647
   */
  Page?: number;
  /**
   * @minimum 1
   * @maximum 2147483647
   */
  PageSize?: number;
};