// src/api/orval/model/openVpnServerWithStatusDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { OpenVpnServerResponse } from "./openVpnServerResponse";
import type { OpenVpnServerStatusLogResponse } from "./openVpnServerStatusLogResponse";

export interface OpenVpnServerWithStatusDto {
  openVpnServerResponses?: OpenVpnServerResponse;
  openVpnServerStatusLogResponse?: OpenVpnServerStatusLogResponse;
  countConnectedClients?: number;
  countSessions?: number;
  totalBytesIn?: number;
  totalBytesOut?: number;
}