// src/api/orval/model/openVpnServerWithStatusResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * OpenVPN Gate Monitor API
 * OpenAPI spec version: v1
 */
import type { OpenVpnServerWithStatusDto } from "./openVpnServerWithStatusDto";

export interface OpenVpnServerWithStatusResponse {
  openVpnServerWithStatus?: OpenVpnServerWithStatusDto;
}